Apply textual decodings to externally-stored attachments
The short-circuit `return` statement in the "external" branch of
_DecodeLOB, introduced in 686daae6, causes externally-stored
attachments to skip the logic that decodes the bytes to characters.
This leads to double-encoded textual content.  This was not
immediately apparent because only textual content longer than the
`$ExternalStorageCutoffSize` cutoff is stored exterally, masking the
problem in most cases.

Remove the premature `return`.  Since the bug was during decoding, not
storage, there is no corrective action necessary for attachments that
were already externalized.

Based on tests and report from Petr <petr@kle.cz>.