Fix problematic case-sensitivity in ExternalAuth::DBI.
Users expect their username and email address to be case-insensitive.
RT's users are case-insensitive, but RT::Authen::ExternalAuth::DBI had a
bug that caused it to fail if its backing database was case-insensitive.

It used DBI::selectall_hashref keyed on the username or email address
from the backing database (whichever was used to look up the account),
and then retrieved the record from the resulting hash based on the value
it had been passed. If the backing database was case-insenitive but
case-preserving (as most are) and the case in the database did not match
the case used for the query, then the case used to create the hash key
would not match the case used to retrieve it. Since Perl hash keys are
case-sensitive, the hash dereference would fail and the method would
return undef, resulting in very strange error messages.

This changes it to use DBI::selectall_arrayref and simply use the first
record returned. Since the code already checks that exactly one result
is returned that change should have no effect on the intended behavior.