Restart asset and catalog ID sequences in upgrade-assets
Without this fix, the asset and catalog ID sequences restart at 1 when you
upgrade from 4.2+Assets to 4.4 on Pg or Oracle. This leads to failures to
create new assets because the id primary key must be unique.

We must acquire an admin dbh because Postgres requires that the owner of the
sequences be the one who resets the sequence.

MySQL and SQLite do not need any adjustment because they do not rely on a
sequence to vend asset/catalog IDs, and so they were unaffected by this bug.

Fixes: T#172082