Avoid spurious "CF changed from 0 to 0" results
Introduced in RT 4.2.9 with 724cf7f845cbdbb2e97c37d008a71ceba4f372a, a
refactoring to address a similar spurious "CF changed from long value to
long value" message (I#30378).

That commit uses the old `($a || '') eq $b` idiom to avoid undef
warnings, but when $a is the value "0", the "eq" operator produces a
false negative, as `(0 || '') eq 0` is evaluated as `'' eq 0`, but the
intent is that it should be true as `0 eq 0`. The modern // operator
avoids the undef warnings but not at the expense of treating "0" in this
broken way.

Fixes: I#32440