For GPG attachment tests on create, use CC, not requestor
If only the requestor is set, only the Autoreply template is used --
which does not set `RT-Attach-Message`, and thus attachments are never
sent.  Because this test is only ever called with `Encrypt => 1`,
however, it only tested that the plaintext attachment content was
lacking -- which it always was, because the attachment was never
involved at all.