Actually include the tested-for text in the attachment
2bf9cb4b refactored the check_text_emails methods out of the two
`gnupg-select-keys-on-___` files.  However, the tests attach $0, the
test file itself -- since the checks are for "Attachment content",
which no longer appears in the test files, this broke the tests.
Since the only use of this string was via `unlike`, it broke them such
that the tests could never fail.

Hardcode the string to search for in the function call, so it exists
in the file that is attached.