Test the unencrypted versions of attachment generation, as well
These tests failed to test what they believed the did because they
only tested the encrypted form; had they tested the unencrypted form,
the lack of the "Attachment content" plaintext, and the lack of
attachments on Autoreply would have been evident.

Add an additional set of tests for enencrypted mails.