Test that expected-encrypted mail is encrypted, with expected contents
Previously, the only test on encrypted content was that it didn't
contain the plaintext.  This fails to verify that it was indeed
encrypted to the right recipient, and that it can successfully be
decrypted to reveal the right content.  Specifically, this makes the
encrypted attachment tests useless, as they cannot prove that the
intended attachment was attached, merely that it was not sent in
plaintext.

Verify that the contents of the outgoing mail can be decrypted, and
that the decrypted text contains the expected string.  This requires
loading not just the public part of the rt-test@example.com user
during tests, but the full secret key as well so that it can be used
to decrypt the data.