Force all DryRun actions to also keep attachments in the session
c8a24b11 switched from using stub methods on RT::Ticket to trying (and
rolling back) the full RT::Interface::Web methods that they were stubs
of.  This means during ticket creation it now simulates links, CFs --
and processing of attachments, which removes them from the session.
As such, when the real processing is later done (or when the creation
page is displayed again, in the case of errors), there are no longer
any attachments.

Whenever used in the context of a DryRun, always keep the attachments.
This is a more general form of f8cbecbc, which only applied to
PreviewScrips; those spot-fixes are no longer necessary.

Fixes: I#32460