Ensure that parse_mail does not GC still-active attachments
Each call to `RT::EmailParser->ParseMIMEEntity*` makes a new
MIME::Parser object, with its own temporary directory (either under
the system tmpdir or under /opt/rt4/var).  These are all removed when
the RT::EmailParser object goes out of scope, via both the tmpdir's
`CLEANUP => 1` as well as via an explicit RT::EmailParser->DESTROY
method.  Thus, if the parsed entities contain attachments. and are not
done being examined by the time their generating RT::EmailParser goes
out of scope, their contents fail to be readable.

Use the same trick that is used in RT::Crypto::GnuPG, where the parsed
MIME::Entity outlives the scope of the RT::EmailParser that generated
it, by storing a reference to the parser inside the generated entity.

This is a stopgap solution; the proper fix requires larger-scale
changes which may be backwards-incompatible.