Add autocomplete for assets
Add support of autocompletion for assets (with RT::Assets-SimpleSearch,
therefore using RT->Config->Get('AssetSearchFields') and
RT->Config->Get('DefaultCatalog')) in field "Add an asset to this
ticket" of the ticket display form.

Also add support of completion both for tickets and for assets in each
input field of the AddLinks form. Completion is done for tickets (as
usual) until the entered term starts with "asset:", where the completion
is switched for assets.

Signed-off-by: gibus <gibus@easter-eggs.com>