Fix bulk update for asset CFs
e6c84c1098d5940ab45a683d1f4197f9e57e7e5f added support for bulk update's
parameter naming convention in order to implement autocomplete CFs on
bulk update. However this change inadvertently broke bulk update of asset CFs
(both in the extension for 4.2 and in core 4.4). Introducing the bulk update
parameter naming convention to ProcessObjectCustomFieldUpdates caused it
to throw errors, as it is not prepared to handle empty $class names.

Since the only caller that needs to handle the bulk update parameter
naming convention is /Helpers/Autocomplete/CustomFieldValues (to parse
the current custom field value out of the form parameters in order to
provide autocomplete suggestions, which was exactly the driver behind
the change in the above commit that introduced the breakage), limiting
_ParseObjectCustomFieldArgs to including the bulk update naming
convention if and only if requested (with the new IncludeBulkUpdate
parameter) maintains the bulk autocomplete feature while restoring asset
CF bulk update.

The loathsome empty capture groups are required because the regular
expression was split into two, made conditional upon the IncludeBulkUpdate
parameter. In the previous factoring of the code, those two capture groups
were in the /^Object-/ branch of the regular expression, so $1 and $2
were empty in the /^Bulk-/ branch.

Fixes: I#32509