Avoid spurious "That is already the current value" for time worked
The mere presence of the argument (which can happen when you submit
Modify.html after having logged time to a ticket) means RT
unconditionally updates the TimeWorked field, which means it
unconditionally adds "That is already the current value" to the results
array.

Instead, only update time worked if the value submitted is not the same
as the current value of TimeWorked on the ticket.

Special consideration is given to setting time worked to the value "0",
which should adjust the time worked on the ticket to 0 rather than being
silently ignored.

This regression was introduced in
1bb578722e7d0a7c57134db82b7ed801cf9d79e8; TimeWorked being an ordinary
field in @attribs means that RT::Record::Update would skip updating that
field because the existing value was the same as the submitted value.
But since TimeWorked no longer uses the Update codepath, there was no
provision to skip it when it's submitted unchanged.