Fix ExternalAuth failure after viewing System Configuration
ExternalSettings's Obfuscate callback was inadvertently overwriting the
password in RT's configuration with "Password not printed", so after
viewing the System Configuration page, any use of ExternalAuth in that
process would fail due to having the incorrect password.

This solution avoids the problem with ExternalSettings specifically, but
also any similar issues we may have elsewhere now and in the future, by
cloning the config data structure before passing it to Obfuscate. That
way any kind of obfuscation is limited in scope to just the System
Configuration page.

Fixes: I#32469