Take a real lock in SetOwner
Ticket updates are done in piecemeal transactions, where individual
pieces, at their own discretion, use transactions to ensure atomicity.
Specifically, `_NewTransaction` and `SetOwner` choose to run within
their own transactions.  The former uses `->LockForUpdate` (see
b99af56f), the latter uses `->_SetLastUpdated` (added in 6061508e).

c8a24b11 changed DryRun transactions (triggered asynchronously from
the ticket update page) to do all of the above steps within a single
database transaction, which is then rolled back.  Because of this
overarching transaction, locks are held across different stages of the
update -- which is a notable difference from the non-DryRun updates.

Together, these can lead to deadlocks when updating owners and
recording correspondence.  The timeline is as follows:

        Ticket/Display.html | PreviewScrips
    ------------------------+----------------------------------------------
     1.                     | Starting transaction in DryRun
                            |
     2. Entering ProcessUpdateMessage
     3. Starting transaction in _NewTransaction
     4. SELECT * FROM Tickets WHERE id = 1 FOR UPDATE
     5. INSERT INTO Transactions ...
     6. UPDATE Tickets SET LastUpdated=... WHERE id=1
     7. Committing transaction
                            |
     8.                     | Entering ProcessUpdateMessage
     9.                     | SELECT * FROM Tickets WHERE id = 1 FOR UPDATE
    10.                     | INSERT INTO Transactions ...
                            |
    11. Entering ProcessTicketBasics
    12. Starting transaction in SetOwner
    13. Doesn't update LastUpdated, as == now already
    14. DELETE FROM CachedGroupMembers WHERE id=...
    15. **BLOCKS**: UPDATE Tickets SET Owner=6 WHERE id=1
                            |
    16.                     | Entering ProcessTicketBasics
    17.                     | **BLOCKS**: DELETE FROM CachedGroupMembers WHERE id=...

Step 9 acquires the lock on the Ticket row for the `PreviewScrips`
request, and step 14 acquires the lock on the `CachedGroupMembers` row
for the update.  The former blocks step 15, the latter blocks step 17.

Note that the ticket update request had previously acquired the lock
on the Ticket row, in step 4, but only for the duration of
`ProcessUpdateMessage`, which it committed in step 7, thus releasing
the lock.  The `PreviewScrips` request, in contrast, ran the same
code, in steps 8-10, but did not release the lock because it was part
of a larger nested transaction.

The "lock" of `->_SetLastUpdated` (introduced in 6061508e) in step 13
is in actuality only functional when there is a race between multiple
different users.  While an update to the Ticket row would indeed serve
to lock it, `->_SetLastUpdated` is only an update if the CurrentUser
and timestamp differ from that stored in the database, as
DBIx::SearchBuilder optimizes the `UPDATE` away if it matches the
current database state.  When the issue is lock contention with
yourself, the `LastUpdated` timestamp already matches from just
previously (step 6), as well as the `LastUpdatedBy`, thus causing no
row update and hence no lock.

Switch to using `->LockForUpdate` instead of `->_SetLastUpdated`.
This takes a reliable lock on the Ticket row; in the above example,
this would block the ticket update at step 13, until PreviewScrips
completed its work and rolled its transaction back.  This suffices to
resolve the new tests in `t/ticket/race.t`.

Fixes: I#32381