Factor out, and use, an atomic transaction method on Tickets
The previous commit introduced a pattern which is likely useful
elsewhere -- that of updating a ticket in a manner that is guaranteed
to be atomic.

Factor out this behavior into a new method, which mirrors DryRun
in some ways.  For additional safety when used with DryRun, the first
SQL statement of both methods is the lock on the Ticket row; this
forces any code within them to run serially.