Allow --debug as a synonym for --verbose, for back-compatibility
75899f244 removed --debug as an option to rt-ldapimport, replacing it
with --verbose.  While --debug had had no effect since it was core'd
in 4.4.0, existing configurations may still contain an invocation with
--debug in their crontabs -- removing it will cause those imports to
begin failing.  Such a change mid-stable-series would be unfortunate.

Allow 'debug', when passed to RT::Interface::CLI's Init, to serve the
same purpose as --verbose.  This allows backwards compatibiity to be
preserved.