Have _SquelchMailTo accept a list of email addresses
This fixes a bug where only the first, rather than every, email address
squelched in a ticket creation transaction was respected. This bug arose
from the discrepency between the documentation and handling of
RT::Ticket->Create's SquelchMailTo, which claims that it accepts a list
of email addresses, and the implementation of
RT::Ticket->_SquelchMailTo, which accepts only a single email address.

Fixes: I#31600