Inline edit for ticket people
This is a slimmed down version of the Modify People page. It does not
support the search form for users and groups, as that requires
reloading the page. It also does not show or edit squelching. For such
advanced needs, Modify People is appropriate.

There are additional measures to make the inline edit panel more svelte,
such as hiding display of multi-user roles with no members, and combining
the tables for single-member roles and multi-member roles.

Fixes: T#178730