When submitting an inline edit form, include all forms' fields
This way if the user edits multiple fields in different inline edit
panels, the page submits all of the fields in each of the panels to the
server.

The usual approach of using a single <form> element surrounding the
entire ticket display page wouldn't work because existing forms (like
create reminders, and create linked ticket in queue) would cause nested
forms, which is unsupported in HTML.

We exclude owner and single-member custom roles from the Basics panel so
that we don't offer multiple form inputs for the same field.

Fixes: T#179970