Avoid Email::Address denial of service attacks
The default configuration of older versions of Email::Address are
vulnerable to algorithmic complexity attacks, creating a denial of
service vulnerability.

Version 1.908 of Email::Address fixes the issue by limiting the
complexity of email addresses it handles. So, this commit bumps the
minimum dependency to that version.

We also set Email::Address::COMMENT_NEST_LEVEL globally to achieve the
same goal, for administrators deploying this as a security patch to
existing RT instances.

This addresses CVE-2015-7686 for RT.

Fixes: T#157608