Don't skip sending mail if there are attached tickets
Ordinarily RT will skip processing an update message (and therefore not
send mail about a comment or correspondence) if the message is empty and
there are no upload attachments. This means we can say the intent of
this code is to allow empty message with just attachments. However, the
code didn't cover the case of attaching tickets instead of files.

This attaching tickets feature is primarily used in RTIR to attach IRs
to mail about an incident.

Fixes: T#180155