Stop double-escaping HTML which is made into links
The MakeClicky component does two things: escapes plain-text content,
and makes inline references to URLs into <a href="..."> links.

It does this by using a regex to find all URL-like things; content
between them is left as-is (if it was originally HTML) or is
HTML-escaped (if it was plaintext).

The replacement, however, is _always_ run on matched content before it
is output.  The result is that a link in HTML with an ampersand, found
as the characters "http://foo/?bar=baz&amp;troz=zort" is HTML-escaped
once again, in both the link and the output, to create:

  <a href="http://foo/?bar=baz&#38;amp;troz=zort">http://foo/?bar=baz&#38;amp;troz=zort</a>

...which is entirely incorrect.

Skip the additional HTML-escaping if the content passed in is already
HTML.  It is not MakeClicky's domain to ensure that HTML is
well-formed and safe -- that is performed by /Elements/ScrubHTML.  As
such, there is no harm in passing through matched content as-is.

Fixes: I#31169