Add Total Time Worked feature
This adds a new recursive TotalTimeWorked method to RT::Ticket
and adds a COLUMN_MAP entry so that it can be displayed in searches.

The feature is optional and requires setting $DisplayTotalTimeWorked
in RT_SiteConfig.pm.

Before creating tickets using this, one needs to disable the 'On
TimeWorked Change Update Parent TimeWorked' scrip which mutates the
value of time worked on parent tickets. It should be noted running
this on an existing database will report invalid results for all tickets
where that scrip was active as it will double-count all of the child
time worked in the parent tickets. This is why the feature is opt in.