adding project time worked
This adds a new recursive project time worked function to Ticket.pm
and adds a COLUMN_MAP entry so that it can be displayed in reports.

The code is optional and requires setting a $DisplayProjectTimeWorked
in RT_SiteConfig.pm.  Also before creating tickets using this one needs
to disable the  'On TimeWorked Change Update Parent TimeWorked' scrip
which mutates the value of time worked on the parent tickets.

It should be noted running this on an existing database will report
invalid results for all tickets where that scrip was active as it will
double count all of the child time worked in the parent tickets.

Fixes: I#181638