While PreviewScrips is reloading, disable its checkboxes
This addresses a race condition between the PreviewScrips AJAX response
coming back and the user checking/unchecking recipients. If the user
timed it incorrectly, then their changes would be lost because it was
too late for the new state to make it to the server, so the response
wouldn't include those changes, and they would be lost when the response
handler replaced the DOM nodes with the server's response.

To prevent users from trying to click on any checkboxes while the
PreviewScrips widget is reloading we disable the checkboxes, and set the
opacity for the panel to 0.3 (to better hint to the user that the UI is
temporarily disabled, since disabled checkboxes can be hard to notice).

Fixes: I#32290