Disable update parent time worked
This patches the initialdata and the upgrade to disable
the scrip which updates parent tickets time worked by
default.  This change allows the new total time worked
feature to be used to roll up the time worked.

The documentation was added to discuss how to renable
this feature if you don't want to use the new feature.