Abstract handle of transaction cf updates on TimeWorked updates
The previous implementation processed any transaction cf updates
in the same transaction as the TimeWorked update. This was added
primarily to support work in the TimeTracking extension. However,
there are use cases where you may want to record transaction cf
updates independent of a TimeWorked update.

Provide a subroutine stub and pass the transaction object so
you can record transaction cf updates against it, but don't make
it the default behavior for all transaction cfs.