Fix QueueCF Charting
When you search using a 'QueueCF.{CF Name}' = 'something', the join will
vector through a hash populated when the packages load with the
LookupTypes for additional classes.  RT::Tickets loads the joins for
RT::Queue and RT::Transaction when the module is loaded.  Because
the join code creates a state hash for each class using it,
RT::Report::Tickets did not have this hash populated.  As a result the
query would always fail.  Adding these join constructors fixes the
search issue, and allows you to search against these lookup types and
chart it.

The test creates 3 queues, and 2 custom fields.  It then performs two
searches, one that aggregates across two of the queues, and one that
only affects a single queue.  Adding the thrid tivial search (no items
returned) is not useful, as it replicates the existing behavior.