Suppress enter-keystroke form submission behavior on ticket create/reply
This is too easy to trigger accidentally, especially when navigating
autocomplete results. This is especially relevant after we added the
drag and drop uploader in 4.4. Prior to 4.4, hitting enter would submit
the form by activating the "Add More Files" button, which of course was
harmless besides just reloading the page. Now in 4.4 it creates/updates
the ticket, which can certainly have irreversible effects.

Fixes: I#32632