make "Name" field and CFs wider for Admin/Groups/Modify.html
We are using some slightly longer group names, however due to the
(unnecessarily) short input field, we can't see the complete value of
what we are typing.

There is plenty of horizontal real estate, so bump the size of the Name
input.

Make the CF size consistent with Name and Description, too.

Remove extraneous (and wrong) 'colspan="3"' for Description table cell.

Reformat table cell tags for readability.