Preload UTF-8 encoding module
After review, Alex pointed out that the root error was Encode::encode
failing to load the encoding due to internal state problems in perl.
This patch pre-loads the encoding before setting up the sig __WARN__
handler.  This will avoid the Encode failing to load.  It should
be noted that Carp.pm too needs to be preloaded or it too can fail
at runtime inside of Encode::encode.

Lastly, because Encode::encode is called without checking enabled,
any UTF-8 strict encoding errors are discarded as warnings are disabled
and the encoding warnings are themselves reissued witha
warnings::warnif. If the encoding warnings need to be debuged the
check paramter Encode::LEAVE_SRC should be passed to Encode::encode.
This will then log the original message.  This might be a useful debug
option.

Fixes: I#32648