Bulk update for SLA
Adds bulk update for SLA. It only adds the field
if there are queues seen in the query results that
have an SLA enabled.  If the SLA update applies to
a ticket in a queue whose SLA is disabled, the SLA
will be updated on the ticket, but will be disabled
at the queue level.  As such the silent success
should be ok.  This means that merely checking the
value of the SLA field itself is insufficient
on the ticket level.