Filter out expired SMIME keys
Without this, we don't show an error in the UI that there is no SMIME
key usable for encryption. Instead we only get an error later when we
finally go to generate encrypted email, at which point it's far too late
to ask the user to decide how to handle the invalid key.

This change mirrors what RT::Crypt::GnuPG::GetKeysForEncryption does
(specifically its Capabilities =~ /e/ check).

Fixes: T#180049