Relax requirements about role names
We can't always have the LookupType readily available inside ValidateName
(e.g. when DBIx::SearchBuilder::Record calls it), and so it's a challenge to
restrict the names to avoid clashing with builtin roles. We could prohibit
any builtin role name, but that means you can't create a "Contact" role for
tickets, nor could you create an AdminCc for assets. So rather than trying
to thread the needle to restrict role names, just open it up.