Support custom roles on asset creation
This required adding a little bit more infrastructure around the Roles()
method passing along arguments to AppliesToObjectPredicate.

The Roles() method lets you pass any arguments, and they will be checked
for the same truthiness as what the role registration provided (e.g.
ACLOnly => 0). However we need to pass the catalog object along because
when you're creating an asset, its ->CatalogObj returns an unloaded
RT::Catalog record. To be able to do so without breaking the truthiness
check, any argument which starts with an underscore is not compared
against the role registration. So we can pass _Catalog from
/Asset/Elements/EditPeople and use it in RT::Asset's
AppliesToObjectPredicate.

This didn't apply to custom roles on tickets because /Ticket/Create.html
iterates over $QueueObj->CustomRoles. Assets instead uses the role system
more generically, so this fix was required.

Similarly, tickets use their queue to resolve roles on create, so switch
assets to use their catalog (instead of the asset itself) to resolve
roles on create.