Fix searching for merged tickets when using MySQL full text index
Fulltext search results would not show tickets that have been merged when searching on a keyword in Content. This makes search work as expected, showing the merge target ticket. The schema for the MySQL full text index table was altered to denormalize and add the tickets' id and EffectiveId. The fulltext tools in sbin are updated to create and maintain the new index.

Fixes T#170768