Check for both id and name mismatches when loading objects
The shredder web interface accepts a name to search for
user objects and up to 4.2.10 this was converted to an ID
in the WipeoutObject list displayed after a search.

f363a0bd changed _AsString to UID for populating these values
which for a user search resulted in usernames being submitted
rather than user ids. Usernames passed to CastObjectsToRecords
loaded properly, but then hit an error check that expected only ids.
Update the check to handle name or id based on the passed value.

Fixes: I#32185