Add Custom Grouping to Admin Group
In order to add the EditCustomFieldCustomGrouping support,
I need to add a default array for the RT::Group.  Trying to
do it in the same fashion as RT::Asset failed due to
require dependencies, so the editing of RT::CustomField
as necessary.

On the Admin/Groups/Modify page I have mirror the layout
of Admin/Users/Modify with the top level table and nested
components.  The layout is single column as the custom field
groupings widget is included as is in a single row.

I confirmed it worked with both variants of the RT_SiteConfig
Set($CustomFieldGroupings, setting with a number of
custom fields and groups.