Calculate time per user dynamically rather than from an attribute
Storing time per user in an attribute saved checking each transaction
on ticket display, however it was not possible to update values
if data ever got out of sync for some reason. In testing, the
performance impact of generating the time dynamically was not
significant, even for tickets with large histories.