Autocompleter for Principals (users and groups)
This is a combination of the autocompleter for users and groups. Both
object types will be searched. The return value is hardcoded: for users,
return their email address or name if email address is empty; for
groups, return "group:Name" (or if "Name" matches any autocomplete
delimiter, "group:ID"). This corresponds with what
RT::Record::Role::Roles::_ResolveRoles accepts.