Reuse CanonicalizePrincipal in DeleteRoleMember
This matches the behavior of AddRoleMember. It's important for followup
commits to be able to support, say, remove "Group:Foo" as AdminCc in
bulk update.

This also fixes a bug where the internal "RT::CustomRole-#" syntax may
show up in error messages when removing a watcher from a custom role
fails.