Add SelfService/Helpers/Upload/Delete
SelfService's attachments dropzone responded correctly to a user
clicking on the button to delete an attachment by removing the
attachment from the dropzone. But RT's JavaScript function to then
delete the attachment from the transaction failed because the path to
/Helpers/Upload/Delete always loaded the version under /. This showed
SelfService users that RT had deleted their attachment when the
resulting transaction would still have the attachment.

This change follows the pattern of /SelfService/Helpers/Upload/Add and
copies /Helpers/Upload/Delete to the
/SelfService/Helpers/Upload/Delete. It also adjusts that paths to
/Helpers/Upload/Delete in RT's attachment deletion JavaScript
functions to use WepHomePath instead of WePath.

Fixes: I#32663