Walk around ACLs when working with lifecycles
This fixes an issue where unprivileged users - essentially anyone
without SeeQueue - would be given the default lifecycle instead of the
queue's lifecycle (because $queue->Lifecycle is guarded by SeeQueue).
This is wholly inappropriate when the queue's lifecycle does not
resemble the default one. The symptom is that users see unexpected
errors like "Status 'custom' isn't a valid status for this ticket" even
when "custom" is a valid status for that ticket's lifecycle, and can be
transitioned to.

So, rather than ignoring the queue's lifecycle, open it up so that when
you're working with a ticket, permissions no longer hide the lifecycle.
This does potentially open up lifecycles that had been previously hidden
by rights, but there are a few mitigating factors. A lifecycle is
unlikely to have private or sensitive information in it. There's also
very little UI for lifecycles beyond the "select status" dropdown.

This mirrors prior art to walk around ACLs to maintain consistency, such
as 68b6a66f.

Fixes: I#32799