Avoid errors in shredder when Organization has a hyphen
Previously, when you shredded RT::Ticket 1 with an $Organization of
"Organ-Ization", then this faulty line of code:

    ($class, $org, $id) = split /-/, $targets;

would leave you with a $class of RT::Ticket, an $org of "Organ", an $id
of "Ization", and the id of 1 being dropped. Since $org "Organ" !=
$Organization "Organ-Ization", you'd get the spurious error:

    Can't wipeout remote object RT::Ticket-Organ-Ization-1

Instead, since what we care about is whether the target is local or
remote by way of checking $Organization, just put $Organization into the
regular expression.

Fixes: I#32406