Avoid errors in shredder when username has a hyphen
Shredder uses username instead of user ID internally (due to
RT::User->UID being overridden to include the username).
But if there's a hyphen in the username, e.g. user-name, then shredder
misinterprets the target "RT::User-user-name" as having an organization of
"user" and an id of "name", leading to this spurious error:

    Can't wipeout remote object RT::User-user-name

RT::User's override of the UID method was introduced without comment in
eb81069b, almost immediately after RT::Record->UID was first added.
However it is difficult to determine whether or not the intent was
_only_ to use the username instead of the id and that Organization was
left off by mistake. Regardless, the new special case here in Shredder
matches the existing special case in RT::User->UID.