Update Sphinx indexer query to match native indexers.
The MySQL and PostgreSQL indexing queries for full-text search have
been updated to include deleted tickets, HTML content, and exclude
content in email transactions. But the Sphinx query that
rt-setup-fulltext-index prints to stdout wasn't similarly updated. The
user must then install the query from rt-setup-fulltext-index. This
change updates the Sphinx query to match the functionality of the
native indexer queries.