Add charts_cache timestamp and clear old cache entries
In typical usage, the charts_cache in the session is cleared
as it is used when a chart is displayed. However, for session stores
that don't support locking, multiple requests can access the session
at the same time and cached charts can end up being written back to
the session after they are deleted.

Add a timestamp for charts_cache entries and a function to clear
entries older than one minute. This doesn't solve the root cause,
but it does provide one more opportunity to clear old cache entries
to prevent them for growing for the lifetime of the session.