Add ExternalAuth to the exceptions for requiring a password
External authentication can be enabled in RT via WebRemoteUserAuth,
which handles authentication in Apache and passes REMOTE_USER to
RT, or via ExternalAuth, which calls an external system from within RT.
Add the second to the configuration checks to determine if
external authentication is enabled and a password check cannot be
done.

Allow an exception for local-only accounts if we detect that
the user account has a password set.