Avoid throwing errors on System Configuration page
The GetObfuscated routine began using Storable::dclone in
cd9bb5215bf935dd53db1b3023d0ee033af0110c to avoid inadvertently updating
the config data structure in memory when obfuscating passwords for the
System Configuration page. That introduced a regression when the
config to be obfuscated included a code reference (which is entirely
reasonable for $ExternalSettings). Storable refuses to serialize code
references, leading to this error when viewing System Configuration:

    Can't store CODE items

This commit fixes the regression by permitting Storable to handle code
references.

Another fix may be to use Storable::forgive_me, but that, according to
the documentation, produces "some meaningless string", which is less
than ideal. Another fix still may be to migrate away from using
Storable::dclone for this purpose, as we are just visualizing.

Fixes: I#33019