Add docs regarding UniqueValues missing column warning
An upgrade step in 4.3.6 creates a new custom field for
CFs that have a category but no CF defined for the BasedOn
field. Creating the new CF before the new CF columns are
added in the 4.4.2 upgrade fails. Add docs with a workaround
since this should impact relatively few users.