Remove AUTHORS from files which don't need it
The AUTHORS information is primarily from extensions that were merged
in, or files dating back to the very first vestiges of the codebase.
It is redundant with the git author information, and the primary
points of contact should not be encouraged to be anything but the
general Discourse site or issues.bestpractical.com

This also removes vestigal and out-of-date information left over from
before when Shredder was merged into core.

Closes GH-233.